pyproject.toml
[project]
name = "personal-finance"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Kien Nguyen", email = "k.nguyen0125@gmail.com"},
]
dependencies = [
    "django>=5.0.6",
    "whitenoise[brotli]>=6.6.0",
    "psycopg[binary]>=3.1.19",
    "djangorestframework>=3.15.1",
    "django-environ>=0.11.2",
    "slippers>=0.6.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
dev = [
    "ruff>=0.4.8",
    "pre-commit>=3.7.1",
]
[tool.pdm]
distribution = false
